Extract some datastructures
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/mtharrison/advent-of-code2024?shareId=XXXX-XXXX-XXXX-XXXX).
