+++ src/day01/mod.rs
@@ -1,18 +1,4 @@
-pub fn parse_input(input: String) -> (Vec<i32>, Vec<i32>) {
-    let mut lhs = Vec::new();
-    let mut rhs = Vec::new();
-
-    for line in input.lines() {
-        let parts: Vec<i32> = line
-            .split_whitespace()
-            .map(|x| x.parse().unwrap())
-            .collect();
-        lhs.push(parts[0]);
-        rhs.push(parts[1]);
-    }
-
-    (lhs, rhs)
-}
+use crate::util::parse_input_day01 as parse_input;
 
 pub fn get_distance(mut lhs: Vec<i32>, mut rhs: Vec<i32>) -> i32 {
     let mut distance = 0;
